<<<<<<< HEAD
\name{ipi.cleanUp}
\alias{ipi.cleanUp}
\title{
ipi.cleanUp
}
\description{
Clears the console and global environment, then unloads packages
}
\usage{
ipi.cleanUp()
}
\author{
Paul I Ighofose
}
=======
\name{ipi.cleanUp}
\alias{ipi.cleanUp}
\title{
ipi.cleanUp
}
\description{
Clears the console and global environment, then unloads packages
}
\usage{
ipi.cleanUp()
}
\author{
Paul I Ighofose
}
>>>>>>> 5863e6f785c1b85928f7351c64dd57201babf032
